♻️ FAQ 도메인 모델 추출 및 데이터 로직 분리
