+++ core/src/main/java/com/linku/core/model/Faq.kt
@@ -1,0 +1,8 @@
+package com.linku.core.model
+
+data class Faq(
+    val id: Int,
+    val question: String,
+    val answer: String,
+    val category: String,
+)
